LayoutTests/accessibility/alternative-presentation-button.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<script src="../resources/js-test.js"></script>
</head>
<body>
<div id="test">
    <h2>First name</h2>
    <input type="text" id="first-name">
    <h2>Last name</h2>
    <input type="text" id="last-name">
</div>
<p id="description"></p>
<div id="console"></div>
<script>
description("This tests that the alternative presentation button is an accessible element and can be hit test.");

if (window.accessibilityController) {
    window.internals.substituteWithAlternativePresentationButton(document.getElementById("test").children, 1);
    var axAlternativePresentationButton = accessibilityController.focusedElement.childAtIndex(0).childAtIndex(0);
    dumpRoleAndDescription(axAlternativePresentationButton);

    debug("<br>Hit test alternative presentation button:");
    var x = axAlternativePresentationButton.clickPointX;
    var y = axAlternativePresentationButton.clickPointY;
    shouldBeTrue("axAlternativePresentationButton.elementAtPoint(x, y).isEqual(axAlternativePresentationButton)");
}

function dumpRoleAndDescription(accessibleElement)
{
    debug("Role: " + accessibleElement.role);
    debug("Description: " + accessibleElement.description);
}
</script>
</body>
</html>
