LayoutTests/accessibility/alternative-presentation-button-expected.txt

This tests that the alternative presentation button is an accessible element and can be hit test.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Role: AXRole: AXButton
Description: AXDescription: alternative presentation button

Hit test alternative presentation button:
PASS axAlternativePresentationButton.elementAtPoint(x, y).isEqual(axAlternativePresentationButton) is true
PASS successfullyParsed is true

TEST COMPLETE
